requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
ollama==0.1.3
crewai==0.1.0
Pillow==10.1.0
python-wordpress-xmlrpc==2.3